+++ packages/brick_offline_first_with_rest/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Allow a class-generic type argument for `OfflineFirstWithRestRepository`
+
 ## 3.2.0
 
 - Add optional `onRequestException` callback function to `RestOfflineQueueClient`